requirements.txt
httpx>=0.28.1
httpx-sse>=0.4.0
jwcrypto>=1.5.6
pydantic>=2.10.6
pyjwt>=2.10.1
sse-starlette>=2.2.1
starlette>=0.46.1
typing-extensions>=4.12.2
uvicorn>=0.34.0
click>=8.1.8
google-adk>=0.0.3
google-genai>=1.9.0
python-dotenv>=1.1.0
crewai[tools]>=0.95.0
langchain-google-genai>=2.0.10
langgraph>=0.3.18
flask
requests
litellm
fastapi
openai
streamlit
